Started separating UILibrary into components
